qa: backtest engine validation + fixes
- Fix look-ahead bias: signals now execute on next bar's open (not same-candle close)
- Fix Sortino NameError: periods_per_year scoped correctly
- Add --days flag to main.py for shorter backtests
- Fix years type hint (int -> float) in data_loader
- Add QA_REPORT.md with full validation results

+++ src/backtest/data_loader.py
@@ -20,7 +20,7 @@ logger = logging.getLogger(__name__)
 def load_binance_ohlcv(
     symbol: str = "BTC/USDT",
     timeframe: str = "15m",
-    years: int = 7,
+    years: float = 7,
     since: Optional[int] = None,
 ) -> pd.DataFrame:
     """Load historical OHLCV data from Binance.
